fix(sample transform): component feature check fix
