+++ src/transforms/mod.rs
@@ -4,6 +4,7 @@ use std::collections::HashSet;
 
 pub mod dedupe;
 pub mod reduce;
+#[cfg(feature = "transforms-impl-sample")]
 pub mod sample;
 
 #[cfg(feature = "transforms-aggregate")]
